api/requirements.txt
flask
python-dotenv
requests
aspose-pdf
aspose-words
